I realized that I have been wrong. I will have to handle preprocessing directives separately from parsing.
Thus, I have added a Haskell module, AbstractSyntaxTree, which defines
types to be parsed. Most of these types are language constructs defined
in N1570 (the final revision of ISO C11). I will eventually create a
typeclass with a function for evaluating types in a constant expression
tree for the preprocessor.