Finished implementing AbstractSyntaxTree module
Will move it around for parts to be reused by compiler proper eventually

In the first commit for this file, I had a C on the end of many data
types to signify that they were for constant expressions. I realized
that this will only matter while evaluating constant expressions, not
while parsing. The C's on the ends of those type variables and
data constructors were pointless and contradict what I plan to do with
those type variables and data constructors in the future.