Started using Data.List.NonEmpty instead of my own NonEmptyList type
Some data types in abstract syntax tree now derive from Show typeclass. This
will aid in debugging. I will probably continue to add a Show
derivation for each data type in the abstract syntax tree as I write a
parser for each.

I added the "defined" unary operator (specific to the preprocessor). See the n1570 pdf.
I implemented it as the "Defined" constructor for the UnaryExpression
data type. I did this because the syntax of the defined operator is much
more similar to sizeof and Alignof unary expressions than other unary
operators.

I filled in the types for each constructor for the GroupPart. I had
forgotten to do that when I first implemented the AST.

A couple name changes for Define Directive constructors in AST.