I put comma parser in its own module.
I put an identifier parser in its own module (distinct from the
identifier CharTokenParser).

I refactored the IdentifierList module. The identifierList parser
(currently the only export), no longer parses for type [String].
Rather, it parses the IdentifierList ADT from the AST.
I am going to make this same transformation in many of the Preprocessor
modules in the near future.

I created a module in the base ccompiler directory called
CustomCombinators. It only exports one object, a custom version of sepBy1,
sepBy1NoConsumption, which I wrote. This function is like sepBy1, except
when sepBy1NoConsumption encounters a trailing delimiter, it does not consume the delimiter
and does not fail.
I put this in its own module because I believe that I will reuse this
for parsing delimited lists in the future. I put this module in the base
compiler directory because I may add to the module in the future.
I will probably move it to a subfolder eventually.