Changes to PreprocessingParser module
I implemented the stringSatisfy_ and stringParserSatisfy_ functions.
I forgot to commit that in the past. Some modules will not build in
previous commits due to me forgetting to add this change.

I changed the type of anyStringToken parser in PreprocessingParser
module to parse String type instead of [String]. In other words,
anyStringToken returns the token type instead of a token stream type.

I wrote a function lexThenParse for debugging purposes. I can lex a
string, then parse the lexed tokens with a parser of type
Show t => PreprocessingParserX t all in one call. This helps give me a
realistic idea of what input text will look like to the parsers for the
preprocessor.

I plan to upscale debugging and testing tools in the coming month or two.
I will probably implement some form of regression testing.