Fixed Define Directive parser in the DefineDirective module
I renamed the defineDirective to defineDirectiveProper.
I then rewrote defineDirective to parse the ControlLine ADT in the
AST, utilizing defineDirectiveProper.

The defineDirectiveProper parser would not consume a newline at the end of a funciton like macro
definition. That newline is required as part of the syntax. See n1570

The DefineDirective module no longer defines the comma parser itself. It
imports comma from the Comma module.

I removed an old comment.

I moved the definePrefix parser to be near the bottom of the file and
out of the way of higher level parsers.