Began moving Preprocessor's Lexer code to use AST
This is opposed to almost exclusively (if not exclusively) parsing
strings from strings in the Lexer.
My plan is to alter the parsers used by the lexer to parse AST data
types that I can reuse or transform to be reused in the compiler proper
and the later stages of the preprocessor.

NOTE: I may have already begun this migration in previous commits, but I
do not think so. I have fallen away from the idea of committing
"atomically" (as I call it). I am not restricting myself to only commit
code that builds because there are many dependencies for me to handle,
and big migrations require many changes that cannot be split into small,
atomic changes.