cleaned up the AbstractSyntaxTree data type definitions, also added some missing ones
The code was only compiling because some of the data constructors were not taking the correct/enough input. Some of those missing inputs were ADTs that were not yet implemented. I do not think that was a matter of me deferring implementation. I think it was a simple mistake of not paying close enough attention.
I also renamed some data constructors.
I made DeclarationSpecifiers much cleaner.